% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wekeo_api.R
\name{download_data}
\alias{download_data}
\title{Title}
\usage{
download_data(hda_list, file_extension = NULL)
}
\arguments{
\item{file_extension}{Optional file extension to add to the downloaded files}

\item{user_filename}{}
}
\value{

}
\description{
Title
}
