% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wekeo_api.R
\name{generate_api_key}
\alias{generate_api_key}
\title{Generate WEkEO API Key}
\usage{
generate_api_key(username, password)
}
\arguments{
\item{username}{: WEkEO username}

\item{password}{: WEkEO password}
}
\value{
a string of the base64-encoded API key
}
\description{
Generates a base64-encoded api key based on the WEkEO user credentials username:password.
}
\details{
Note that decoding the hash will recover the original username and password. For that reason it is strongly advised that you never include the API key in materials which are publically shared.
}
