% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wekeo_api.R
\name{get_access_token}
\alias{get_access_token}
\title{Title}
\usage{
get_access_token(hda_list)
}
\arguments{
\item{hda_list}{: list initiated with the function init}
}
\value{
the list including the access token
}
\description{
Requests an access token to use the HDA API and stores it as a separate key in the dictionary
}
