% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wekeo_api.R
\name{get_order_ids}
\alias{get_order_ids}
\title{get_order_ids}
\usage{
get_order_ids(hda_list)
}
\arguments{
\item{hda_list}{}
}
\value{
hda_list including order ids and the request staus of assigning the order Ids.
}
\description{
Gets the set of order ids corresponding to query results
}
