% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wekeo_api.R
\name{get_query_results_list}
\alias{get_query_results_list}
\title{get_query_results_list}
\usage{
get_query_results_list(hda_list)
}
\arguments{
\item{hda_list}{}
}
\value{

}
\description{
get_query_results_list
}
