% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wekeo_api.R
\name{generate_api_key}
\alias{generate_api_key}
\title{generate_api_key}
\usage{
generate_api_key(username, password)
}
\arguments{
\item{username}{: WEkEO username}

\item{password}{: WEkEO password}
}
\value{
a string of the base64-encoded API key
}
\description{
Generates a base64-encoded api key based on the WEkEO user credentials username:password.
}
